Relocates content into feature package.
