+++ financial-connections/src/main/java/com/stripe/android/financialconnections/di/FinancialConnectionsSubcomponentBuilderProvider.kt
@@ -1,6 +1,7 @@
 package com.stripe.android.financialconnections.di
 
 import com.airbnb.mvrx.ViewModelContext
+import com.stripe.android.financialconnections.features.institutionpicker.InstitutionPickerSubcomponent
 import com.stripe.android.financialconnections.ui.FinancialConnectionsSheetNativeActivity
 import javax.inject.Inject
 import javax.inject.Provider
